+++ GymAPI/Models/Client/Client.cs
@@ -12,7 +12,7 @@ namespace GymAPI.Models
         public string ImageUrl { get; set; }
         public DateTime BirthDate { get; set; }
         public int Age { get; set; }
-        public float HeightInMeters { get; set; }
+        public double HeightInMeters { get; set; }
         public float WeightInKg { get; set; }
         public TrainingPlan TrainingPlan { get; set; }
